services/rides/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
